consolidated the set state commands into one
