+++ Team-3749-2025/simgui-ds.json
@@ -90,7 +90,6 @@
     }
   ],
   "robotJoysticks": [
-    {},
     {
       "guid": "Keyboard0"
     }
