--- Team-3749-2025/src/main/java/frc/robot/commands/arm/SetCoralState.java
@@ -1,40 +1,0 @@
-package frc.robot.commands.arm;
-
-import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.arm.coral.CoralConstants.ArmStates;
-import frc.robot.subsystems.arm.coral.CoralArm;
-import frc.robot.utils.UtilityFunctions;
-
-/**
- * Sets the selected state for a selected coral arm
- * @author Weston Gardner
- */
-
-public class SetCoralState extends Command {
-
-    CoralArm arm;
-    ArmStates state;
-    double setPoint;
-
-    public SetCoralState(CoralArm arm, ArmStates state, double setPoint) {
-        this.setPoint = setPoint;
-        this.state = state;
-        this.arm = arm;
-        super.addRequirements(arm);
-    }
-
-    @Override
-    public void initialize() {
-        arm.setState(state);
-    }
-
-    @Override
-    public void end(boolean interupted) {
-    }
-
-    @Override
-    public boolean isFinished() {
-        double position = arm.getPositionRad();
-        return UtilityFunctions.withinMargin(0.0001, setPoint, position);
-    }
-}
